Minor improvement
Use System.out instead of System.err, to be sure that the lines are displayed at the right place. The Eclipse IDE seems to sometimes mix the two, so a Sysout may be displayed after a Syserr, even if the Syserr was done BEFORE in the code. Yes, that's weird!